permanencia_modelos/requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
python-dotenv>=1.0
matplotlib>=3.8

catboost>=1.2
lightgbm>=4.3
xgboost>=2.0
shap>=0.45
